+++ tests/modules/lint/test_module_lint_remotes.py
@@ -23,8 +23,8 @@ class TestModulesLintRemotes(TestModules):
 
     def test_modules_lint_gitlab_modules(self):
         """Lint modules from a different remote"""
-        self.mods_install_gitlab.install("fastqc")
-        self.mods_install_gitlab.install("multiqc")
+        self.mods_install_gitlab_nftest.install("fastqc")
+        self.mods_install_gitlab_nftest.install("multiqc")
         module_lint = nf_core.modules.lint.ModuleLint(directory=self.pipeline_dir, remote_url=GITLAB_URL)
         module_lint.lint(print_results=False, all_modules=True)
         assert len(module_lint.failed) == 3
@@ -33,9 +33,9 @@ class TestModulesLintRemotes(TestModules):
 
     def test_modules_lint_multiple_remotes(self):
         """Lint modules from a different remote"""
-        self.mods_install_gitlab.install("multiqc")
+        self.mods_install_gitlab_nftest.install("multiqc")
         module_lint = nf_core.modules.lint.ModuleLint(directory=self.pipeline_dir, remote_url=GITLAB_URL)
-        module_lint.lint(print_results=False, all_modules=True)
-        assert len(module_lint.failed) == 1
+        module_lint.lint(print_results=True, all_modules=True)
+        assert len(module_lint.failed) == 2
         assert len(module_lint.passed) > 0
         assert len(module_lint.warned) >= 0